+++ event-sourcing/event-publisher/event-publisher-timer/src/test/java/uk/gov/justice/services/eventsourcing/publishedevent/prepublish/EventPublishIT.java
@@ -199,7 +199,7 @@ public class EventPublishIT {
             OversizeMessageGuard.class,
             JmsMessagingConfiguration.class,
             ValueProducer.class,
-            EventLinkingNotifier.class,
+            PrePublishNotifier.class,
             EventPublishingNotifier.class
     })
     public WebApp war() {